Add util for encodeAndAddSoundToVM
