+++ src/lib/audio/audio-util.js
@@ -1,3 +1,5 @@
+import WavEncoder from 'wav-encoder';
+
 const SOUND_BYTE_LIMIT = 10 * 1000 * 1000; // 10mb
 
 const computeRMS = function (samples, scaling = 0.55) {
@@ -23,8 +25,43 @@ const computeChunkedRMS = function (samples, chunkSize = 1024) {
     return chunkLevels;
 };
 
+const encodeAndAddSoundToVM = function (vm, samples, sampleRate, name, callback) {
+    WavEncoder.encode({
+        sampleRate: sampleRate,
+        channelData: [samples]
+    }).then(wavBuffer => {
+        const vmSound = {
+            format: '',
+            dataFormat: 'wav',
+            rate: sampleRate,
+            sampleCount: samples.length
+        };
+
+        // Create an asset from the encoded .wav and get resulting md5
+        const storage = vm.runtime.storage;
+        vmSound.asset = storage.createAsset(
+            storage.AssetType.Sound,
+            storage.DataFormat.WAV,
+            new Uint8Array(wavBuffer),
+            null,
+            true // generate md5
+        );
+        vmSound.assetId = vmSound.asset.assetId;
+
+        // update vmSound object with md5 property
+        vmSound.md5 = `${vmSound.assetId}.${vmSound.dataFormat}`;
+        // The VM will update the sound name to a fresh name
+        vmSound.name = name;
+
+        vm.addSound(vmSound).then(() => {
+            if (callback) callback();
+        });
+    });
+};
+
 export {
     computeRMS,
     computeChunkedRMS,
+    encodeAndAddSoundToVM,
     SOUND_BYTE_LIMIT
 };
